basic input/output example
NOTE FROM VIDEO:

The GREEN    //    are comments. The compiler ignores comments. I'll show you how to make a comment block later. Comments are used to keep your code clear and concise. They're for you and other programmers. 

Variables do NOT only have to start with a lowercase letter followed by an uppercase letter. In the video, I used faveColor. You can also do this ex: int my_Name. Notice how the second word is still capitalized. This is a style preference. You can also do int name or int first_name. Another example would be int my_name. When it comes to variables, there are additional rules you will learn. 

However, it is good to know there are many styles that are approved and can be used.

Remember you need to create a _variable_ with the data type in order to store the information. If you ask a question before creating a place to put it, then where is going to go? These are the questions to ask yourself.